replace Data with DataSet
