+++ src/main/java/ru/naumen/perfhouse/influx/InfluxDAO.java
@@ -39,10 +39,10 @@ import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Component;
 
 import ru.naumen.perfhouse.statdata.Constants;
-import ru.naumen.sd40.log.parser.data.ActionDoneData;
-import ru.naumen.sd40.log.parser.data.ErrorData;
-import ru.naumen.sd40.log.parser.data.GCData;
-import ru.naumen.sd40.log.parser.data.TopData;
+import ru.naumen.sd40.log.parser.dataset.ActionDoneDataSet;
+import ru.naumen.sd40.log.parser.dataset.ErrorDataSet;
+import ru.naumen.sd40.log.parser.dataset.GCDataSet;
+import ru.naumen.sd40.log.parser.dataset.TopDataSet;
 
 /**
  * Created by doki on 24.10.16.
@@ -107,8 +107,8 @@ public class InfluxDAO
         return BatchPoints.database(dbName).build();
     }
 
-    public void storeActionsFromLog(BatchPoints batch, String dbName, long date, ActionDoneData dones,
-            ErrorData errors)
+    public void storeActionsFromLog(BatchPoints batch, String dbName, long date, ActionDoneDataSet dones,
+            ErrorDataSet errors)
     {
         //@formatter:off
         Builder builder = Point.measurement(Constants.MEASUREMENT_NAME).time(date, TimeUnit.MILLISECONDS)
@@ -176,7 +176,7 @@ public class InfluxDAO
         }
     }
 
-    public void storeGc(BatchPoints batch, String dbName, long date, GCData gc)
+    public void storeGc(BatchPoints batch, String dbName, long date, GCDataSet gc)
     {
         Point point = Point.measurement(Constants.MEASUREMENT_NAME).time(date, TimeUnit.MILLISECONDS)
                 .addField(GCTIMES, gc.getGcTimes()).addField(AVARAGE_GC_TIME, gc.getCalculatedAvg())
@@ -192,7 +192,7 @@ public class InfluxDAO
         }
     }
 
-    public void storeTop(BatchPoints batch, String dbName, long date, TopData data)
+    public void storeTop(BatchPoints batch, String dbName, long date, TopDataSet data)
     {
         Point point = Point.measurement(Constants.MEASUREMENT_NAME).time(date, TimeUnit.MILLISECONDS)
                 .addField(AVG_LA, data.getAvgLa()).addField(AVG_CPU, data.getAvgCpuUsage())
