+++ src/main/java/ru/naumen/perfhouse/uploaders/GCUploader.java
@@ -1,6 +1,5 @@
 package ru.naumen.perfhouse.uploaders;
 
-import ru.naumen.sd40.log.parser.data.GCData;
 import ru.naumen.sd40.log.parser.dataset.GCDataSet;
 
 public class GCUploader extends DBUploader<GCDataSet>{
@@ -10,10 +9,9 @@ public class GCUploader extends DBUploader<GCDataSet>{
 
     @Override
     public void upload(Long key, GCDataSet dataSet) {
-        GCData gc = dataSet.getGcData();
-        if (!gc.isNan())
+        if (!dataSet.isNan())
         {
-            storage.storeGc(batchPoints, influxDb, key, gc);
+            storage.storeGc(batchPoints, influxDb, key, dataSet);
         }
     }
 }
