--- src/main/java/ru/naumen/sd40/log/parser/data/GCData.java
@@ -1,34 +1,0 @@
-package ru.naumen.sd40.log.parser.data;
-
-import org.apache.commons.math3.stat.descriptive.DescriptiveStatistics;
-
-import static ru.naumen.sd40.log.parser.NumberUtils.getSafeDouble;
-import static ru.naumen.sd40.log.parser.NumberUtils.roundToTwoPlaces;
-
-public class GCData implements ParserData{
-    private DescriptiveStatistics ds = new DescriptiveStatistics();
-
-    public double getCalculatedAvg()
-    {
-        return roundToTwoPlaces(getSafeDouble(ds.getMean()));
-    }
-
-    public long getGcTimes()
-    {
-        return ds.getN();
-    }
-
-    public double getMaxGcTime()
-    {
-        return roundToTwoPlaces(getSafeDouble(ds.getMax()));
-    }
-
-    public boolean isNan()
-    {
-        return getGcTimes() == 0;
-    }
-
-    public void addValue(Double value){
-        ds.addValue(value);
-    }
-}
