--- src/main/java/ru/naumen/sd40/log/parser/data/ParserData.java
@@ -1,4 +1,0 @@
-package ru.naumen.sd40.log.parser.data;
-
-public interface ParserData {
-}
