--- src/main/java/ru/naumen/sd40/log/parser/data/TopData.java
@@ -1,62 +1,0 @@
-package ru.naumen.sd40.log.parser.data;
-
-import org.apache.commons.math3.stat.descriptive.DescriptiveStatistics;
-
-import static ru.naumen.sd40.log.parser.NumberUtils.getSafeDouble;
-import static ru.naumen.sd40.log.parser.NumberUtils.roundToTwoPlaces;
-
-public class TopData implements ParserData{
-    private DescriptiveStatistics laStat = new DescriptiveStatistics();
-    private DescriptiveStatistics cpuStat = new DescriptiveStatistics();
-    private DescriptiveStatistics memStat = new DescriptiveStatistics();
-
-    public void addLa(double la)
-    {
-        laStat.addValue(la);
-    }
-
-    public void addCpu(double cpu)
-    {
-        cpuStat.addValue(cpu);
-    }
-
-    public void addMem(double mem)
-    {
-        memStat.addValue(mem);
-    }
-
-    public boolean isNan()
-    {
-        return laStat.getN() == 0 && cpuStat.getN() == 0 && memStat.getN() == 0;
-    }
-
-    public double getAvgLa()
-    {
-        return roundToTwoPlaces(getSafeDouble(laStat.getMean()));
-    }
-
-    public double getAvgCpuUsage()
-    {
-        return roundToTwoPlaces(getSafeDouble(cpuStat.getMean()));
-    }
-
-    public double getAvgMemUsage()
-    {
-        return roundToTwoPlaces(getSafeDouble(memStat.getMean()));
-    }
-
-    public double getMaxLa()
-    {
-        return roundToTwoPlaces(getSafeDouble(laStat.getMax()));
-    }
-
-    public double getMaxCpu()
-    {
-        return roundToTwoPlaces(getSafeDouble(cpuStat.getMax()));
-    }
-
-    public double getMaxMem()
-    {
-        return roundToTwoPlaces(getSafeDouble(memStat.getMax()));
-    }
-}
